configure sort command
