+++ src/extension.ts
@@ -5,9 +5,11 @@ import { configureDiagnostics } from './diagnostics'
 import { configureGenerator } from './generator'
 import { configureReferenceProvider } from './reference'
 import { configureSchemaStatus } from './schema'
+import { configureSortSchemaCommand } from './sort'
 
 export async function activate(context: vscode.ExtensionContext) {
   configureConfigWatcher(context)
+  configureSortSchemaCommand(context)
   await configureSchemaStatus(context)
   await configureDiagnostics(context)
   await configureGenerator(context)
